refactor: Move fees functions to utils file
