+++ src/components/Fees.tsx
@@ -1,5 +1,4 @@
 import { BigNumber } from "bignumber.js";
-import type { Accessor } from "solid-js";
 import {
     Show,
     createEffect,
@@ -10,7 +9,6 @@ import {
 } from "solid-js";
 
 import { config } from "../config";
-import { LBTC } from "../consts/Assets";
 import { SwapType } from "../consts/Enums";
 import { useCreateContext } from "../context/Create";
 import { useGlobalContext } from "../context/Global";
@@ -24,18 +22,14 @@ import {
     calculateBoltzFeeOnSend,
     calculateSendAmount,
 } from "../utils/calculate";
-import { isConfidentialAddress } from "../utils/compat";
 import { formatAmount } from "../utils/denomination";
+import { isToUnconfidentialLiquid, unconfidentialExtra } from "../utils/fees";
 import { getPair } from "../utils/helper";
 import { weiToSatoshi } from "../utils/rootstock";
 import { getClaimAddress } from "./CreateButton";
 import Denomination from "./settings/Denomination";
 
 const ppmFactor = 10_000;
-
-// When sending to an unconfidential address, we need to add an extra
-// confidential OP_RETURN output with 1 sat inside
-export const unconfidentialExtra = 5;
 
 const rifExtraGasCost = 157_000n;
 
@@ -50,19 +44,6 @@ export const getFeeHighlightClass = (fee: number, regularFee: number) => {
 
     return "";
 };
-
-export const isToUnconfidentialLiquid = ({
-    assetReceive,
-    addressValid,
-    onchainAddress,
-}: {
-    assetReceive: Accessor<string>;
-    addressValid: Accessor<boolean>;
-    onchainAddress: Accessor<string>;
-}) =>
-    assetReceive() === LBTC &&
-    addressValid() &&
-    !isConfidentialAddress(onchainAddress());
 
 const Fees = () => {
     const {
